ai/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.110.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.0
fsspec==2025.3.2
future==1.0.0
h11==0.14.0
huggingface-hub==0.30.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==24.2
pydantic==2.11.1
pydantic_core==2.33.0
pydub==0.25.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.1.99
setuptools==78.1.0
sniffio==1.3.1
SpeechRecognition==3.14.1
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.36.3
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.15.2
torch==2.6.0
tqdm==4.67.1
transformers==4.36.1
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.29.0
Werkzeug==3.1.3
wheel==0.45.1